halium: boot into recovery on A/B devices by including it into halium-ramdisk
The script will look for /ramdisk-recovery.img in the root of ramdisk
and assume device is A/B if it exists. Then, if bootloader does not
put "skip_initramfs" or "androidboot.force_normal_boot=1" in cmdline,
it triggers boot into recovery.

Cherry-picked from Volla Phone 22 [1] with dropped unnecessary
/sbin/init creation that also breaks creating the usual symlink and
avoiding extra cat -> gzip pipeline.

[1] https://gitlab.com/ubports/porting/reference-device-ports/android11/volla-phone-22/volla-mimameid/-/commit/882ee97b)